Adjust file names (cocktails / mocktails) adjust styling for responsiveness and change background color
