--- src/components/FavoriteDrinksGrid/FavoriteDrinksGrid.css
@@ -1,54 +1,0 @@
-.drinks-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
-  gap: 3em;
-  padding: 3em;
-  justify-items: center;
-
-  .drink-card {
-    width: 220px;
-    height: 350px;
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    align-items: center;
-    border: solid 1px #333;
-    padding: .3em;
-    background: rgb(22, 22, 22) content-box;
-    box-shadow: 0 0 10px rgba(0, 0, 0, 0.2);
-    overflow: hidden;
-    transition: transform 0.3s;
-    color: antiquewhite; 
-    &:hover {
-      transform: scale(1.05);
-    }
-    h2 {
-      font-size: large;
-      margin: 0.5em;
-    }
-
-  }
-}
-
-.heart-icon {
-  height: 55%;
-}
-
-.drink-card-info {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 20%;
-  margin-left: 2%;
-  margin-right: 2%;
-}
-
-.drink-card-image {
-  margin-top: 5%;
-  height: 80%;
-  width: 95%;
-}
-
-a {
-  text-decoration: none;
-}
