+++ src/components/HomePage/HomePage.css
@@ -47,3 +47,11 @@
     
   }
 }
+
+
+@media (max-width: 700px) {
+  .App {
+   transform: scale(0.8);
+   height: 10vh;
+ }
+}
